Import GtkColorSelection as Gcolor3ColorSelection
We're going to have to modify it to make it work with any upcoming
Wayland solution, and it has been removed from upstream master.

This imported version has some deprecated API updated (more to be done)
and is modified to compile with our stricter compiler flags. There are
some other cleanups to be done, too (e.g. we can probably remove the
separate alpha property now that it uses GdkRGBA).