Gcolor3ColorStore: implement GListModel
By implementing the GListModel interface, we can bind our (upcoming)
GtkListBox directly to the store, so every change to the store is
automatically reflected in the listbox, without us having to do
anything.

In the future, we might want to switch away from having a GKeyFile as
backend. Right now, we need to iterate it linearly every time we add,
remove or rename a color. I doubt that users will save millions of
colors in Gcolor3, so performance should be OK, but something more
performant is always better.

Also, we are limited in our renaming options due to GKeyFile. See the
comment in the code.